requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
bleach==1.5.0
Brlapi==0.6.4
certifi==2018.4.16
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.36.0
django-password-reset==1.0
django-recaptcha==1.4.0
django-recaptcha2==1.0.3
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-swagger==2.0.7
django-secure==1.0.1
django-ssl-auth==2.0.1
django-sslserver==0.20
django-tinymce==2.7.0
djangorestframework==3.8.2
entrypoints==0.2.3
enum34==1.1.6
gunicorn==19.8.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.3.2.1
Keras==2.1.2
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.3
mysql-connector==2.1.6
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.14.0
oauthlib==2.0.7
openapi-codec==1.3.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.6.1
pyparsing==2.2.0
python-apt===1.1.0.b1-elementary2-r2-ubuntu0.4.1.1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
python3-openid==3.1.0
pytz==2018.4
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
reportlab==3.3.0
requests==2.18.4
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.0.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.14.0
six==1.11.0
sklearn==0.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
system-service==0.3
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.22
usb-creator==0.3.0
uWSGI==2.0.17
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
xdiagnose==3.8.4.1
xkit==0.0.0
